requirements.txt

eth-brownie==1.6.8
